requirements.txt
# Core Dependencies
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2

# Feature Store & ML Platform
hopsworks==3.4.4

# API Requests
requests==2.31.0
python-dotenv==1.0.0

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Web Dashboard
streamlit==1.29.0
gradio==4.10.0

# Model Explainability
shap==0.44.0

# Deep Learning (Optional)
tensorflow==2.15.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3